fix migration

+++ Gym.Models/InvoiceDetail.cs
@@ -9,9 +9,9 @@ namespace Gym.Models
 	public class InvoiceDetail
 	{
 		public int InvoiceDetailId { get; set; }
-		public decimal PricePerUnit { get; set; }
+		public int PricePerUnit { get; set; }
 		public int Quantity { get; set; }
-		public decimal Amount { get; set; }
+		public int Amount { get; set; }
 
 		public Invoice Invoice { get; set; }
 		public Service Service { get; set; }
